+++ opentelemetry-sdk/src/metrics/meter.rs
@@ -395,7 +395,7 @@ impl SdkMeter {
         }
 
         if let Some(ref boundaries) = builder.boundaries {
-            let validation_result = validate_buckets(boundaries);
+            let validation_result = validate_bucket_boundaries(boundaries);
             if let Err(err) = validation_result {
                 // TODO: Include the buckets too in the error message.
                 // TODO: This validation is not done when Views are used to
@@ -550,16 +550,19 @@ fn validate_instrument_config(name: &str, unit: &Option<Cow<'static, str>>) -> M
     validate_instrument_name(name).and_then(|_| validate_instrument_unit(unit))
 }
 
-fn validate_buckets(buckets: &[f64]) -> MetricResult<()> {
-    if buckets.is_empty() {
-        return Err(MetricError::InvalidInstrumentConfiguration(
-            "Buckets must not be empty",
-        ));
+fn validate_bucket_boundaries(boundaries: &[f64]) -> MetricResult<()> {
+    // Validate boundaries do not contain f64::NAN, f64::INFINITY, or f64::NEG_INFINITY
+    for boundary in boundaries {
+        if boundary.is_nan() || boundary.is_infinite() {
+            return Err(MetricError::InvalidInstrumentConfiguration(
+                "Buckets must not contain NaN, +Inf, or -Inf",
+            ));
+        }
     }
 
     // validate that buckets are sorted and non-duplicate
-    for window in buckets.windows(2) {
-        if window[0] >= window[1] {
+    for i in 1..boundaries.len() {
+        if boundaries[i] <= boundaries[i - 1] {
             return Err(MetricError::InvalidInstrumentConfiguration(
                 "Buckets must be sorted and non-duplicate",
             ));