+++ opentelemetry-sdk/CHANGELOG.md
@@ -18,6 +18,10 @@
         .build();
     ```
 
+  - Bug fix: Empty Logger names are retained as-is instead of replacing with
+    "rust.opentelemetry.io/sdk/logger"
+    [#2334](https://github.com/open-telemetry/opentelemetry-rust/pull/2316)
+
 ## 0.27.0
 
 Released 2024-Nov-11