+++ opentelemetry-sdk/CHANGELOG.md
@@ -33,9 +33,10 @@
 
 
   - **Bug Fix:** Validates the `with_boundaries` bucket boundaries used in
-    Histograms. The boundaries provided by the user must be a non-empty vector,
-    sorted in strictly increasing order, and contain no duplicates. Instruments
-    will not record measurements if the boundaries are invalid.
+    Histograms. The boundaries provided by the user must not contain `f64::NAN`,
+    `f64::INFINITY`, `f64::NEG_INFINITY` and must be sorted in strictly
+    increasing order, and contain no duplicates. Instruments will not record
+    measurements if the boundaries are invalid.
     [#2351](https://github.com/open-telemetry/opentelemetry-rust/pull/2351)
 
 ## 0.27.0